// server/src/types.d.ts
export declare enum FactionId {
    REPUBLICANS = "REPUBLICANS",
    CONSPIRATORS = "CONSPIRATORS",
    NOBLES = "NOBLES",
    NEUTRAL = "NEUTRAL",
    LOYALISTS = "LOYALISTS",
    PRINCELY_ARMY = "PRINCELY_ARMY",
    CONFEDERATE_CITIES = "CONFEDERATE_CITIES"
}
export interface PlayerInfo {
    odId: string;
    faction: FactionId | null;
    isHost: boolean;
    isReady: boolean;
    isConnected: boolean;
    nickname?: string;
}